requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
kiwisolver==1.4.4
matplotlib==3.7.1
mrcfile==1.4.3
numpy==1.24.2
packaging==23.0
Pillow==9.5.0
progress==1.6
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
